requirements.txt
opencv-python-headless==4.10.0.84
pika==1.3.2
keras==3.3.2
surrealdb==0.3.2
mtcnn==0.1.1
keras_vggface @ git+https://github.com/findursmile/keras-vggface.git@81a62f1694294ed716406fe24e81dd45d7be29cc
Keras-Applications==1.0.8
matplotlib==3.8.4
python-dotenv==1.0.1
python-telegram-bot==21.6
